pyproject.toml
[project]
name = "office-ocr-rpa-template"
version = "0.1.0"
description = "Screen OCR, coordinate clicking, and data-driven office automation workflow templates."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "rapidocr>=3.0.0",
  "onnxruntime>=1.17.0",
  "mss>=9.0.1",
  "Pillow>=10.0.0",
  "pyautogui>=0.9.54",
  "keyboard>=0.13.5",
  "PyYAML>=6.0.1",
  "openpyxl>=3.1.2",
  "rapidfuzz>=3.6.1",
  "pyperclip>=1.8.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
